chore: update tsconfig for e2e
