+++ e2e/tests/envEnable.spec.ts
@@ -14,7 +14,7 @@ import {
     openTestingView,
     waitForResolved,
     waitForTestStatus,
-} from '../helpers/index.ts'
+} from '../helpers/index.js'
 
 import type { SideBarView, Workbench } from 'wdio-vscode-service'
