+++ e2e/tsconfig.json
@@ -6,12 +6,12 @@
     "types": ["node", "@wdio/globals/types", "@wdio/mocha-framework", "wdio-vscode-service"],
     "skipLibCheck": true,
     "noEmit": true,
-    "allowImportingTsExtensions": true,
     "resolveJsonModule": true,
     "isolatedModules": true,
     "strict": true,
     "noUnusedLocals": true,
     "noUnusedParameters": true,
+    "experimentalDecorators": true,
     "noFallthroughCasesInSwitch": true
   }
 }
